Closes #2 (again)
I tried to add a having clause to the statement several times but SQL kept throwing a syntax error, which is why I tried coalesce in the first place. When reading the laravel docs I found a function named havingRaw() which works fine when you are trying to use an "is not null" statement. trying to use having('field','is not null') will produce a broken query.